backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
chromadb
pypdf
python-dotenv
duckduckgo-search
sentence-transformers
